PostgreSQLのserial→bigserial移行とJavaのInteger→Long変換
将来的なデータ増加に備え、PKをserial→bigserial、FK/参照カラムをint→bigintに変更し、
Java側のID関連フィールド(accountNo, photoNo, tagNo, locationNo等)をInteger→Longに統一する。

+++ backend/src/main/java/com/web/gallery/AccountPrincipal.java
@@ -43,7 +43,7 @@ public class AccountPrincipal implements UserDetails {
 	 *
 	 * @return	アカウント番号
 	 */
-	public Integer getAccountNo() {
+	public Long getAccountNo() {
 		return accountModel.getAccountNo();
 	}
