+++ backend/src/main/java/com/web/gallery/dto/PhotoDetailDto.java
@@ -13,10 +13,10 @@ import lombok.Data;
 @Data
 public class PhotoDetailDto {
 	/** アカウントNo */
-	private Integer accountNo;
-	
+	private Long accountNo;
+
 	/** 写真番号 */
-	private Integer photoNo;
+	private Long photoNo;
 
 	/** お気に入り */
 	private Boolean isFavorite;
@@ -25,7 +25,7 @@ public class PhotoDetailDto {
 	private OffsetDateTime photoAt;
 	
 	/** ロケーション番号 */
-	private Integer locationNo;
+	private Long locationNo;
 	
 	/** 住所 */
 	private String address;
