+++ db/common/location_mst.sql
@@ -6,17 +6,17 @@ DROP TABLE IF EXISTS common.location_mst;
 CREATE TABLE common.location_mst
 (
 	-- ID
-	id serial NOT NULL,
+	id bigserial NOT NULL,
 	-- アカウント番号
-	account_no int NOT NULL,
+	account_no bigint NOT NULL,
 	-- ロケーション番号
-	location_no int NOT NULL,
+	location_no bigint NOT NULL,
 	-- 作成者
-	created_by int NOT NULL,
+	created_by bigint NOT NULL,
 	-- 作成日時
 	created_at timestamp with time zone NOT NULL,
 	-- 更新者
-	updated_by int NOT NULL,
+	updated_by bigint NOT NULL,
 	-- 更新日時
 	updated_at timestamp with time zone NOT NULL,
 	-- 削除フラグ
